refactor: Split index.js into small modules
No functionality was changed here, just splitting it up to make it easy to work on.
