+++ lib/test/prepare-request.test.js
@@ -1,0 +1,10 @@
+const prepareRequest = require('../prepare-request')
+
+describe('prepareRequest', () => {
+  test('removes options.overwrite when false', () => {
+    expect(prepareRequest({ overwrite: false }).formData).toEqual({})
+  })
+  test('does not remove options.overwrite when true', () => {
+    expect(prepareRequest({ overwrite: true }).formData).toEqual({ overwrite: 'true' })
+  })
+})
